+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ dev
 ---
 
 - Furhther performance improvements for rerunning inference using stored data via caches
+- Added the general Gaussian noise example model (fixed covariance)
 
 0.6.2 (2017-09-06)
 ------------------